pow functions return number of type long

+++ Sem.04/Pract.04/09.cpp
@@ -19,7 +19,7 @@ int getDigitsCount(int n)
     return count;
 }
 
-int pow(int n, int k)
+long pow(long n, int k)
 {
     int result = 1;
